backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
celery[redis]==5.4.0
redis==5.2.1
httpx==0.28.1
pillow==11.1.0
python-multipart==0.0.19
websockets==14.1
python-dotenv==1.0.1
